[Feat] User 엔티티 introduction 컬럼 추가

+++ keewe-domain/src/main/java/ccc/keewedomain/service/user/ProfileDomainService.java
@@ -200,6 +200,8 @@ public class ProfileDomainService {
 
     private void deleteProfilePhoto(User user) {
         user.deleteProfilePhoto();
-        storeService.delete(user.getProfilePhotoURL());
+        if(user.getProfilePhoto() != null ) {
+            storeService.delete(user.getProfilePhotoURL());
+        }
     }
 }
